fix: resolve types of loop iterators

+++ src/CorrelatorTimeSeries.cpp
@@ -159,7 +159,7 @@ PointCollection CorrelatorTimeSeries::filterPointsInRange(TimeRange range) {
       accumulator_set<double, stats<tag::covariance<double, tag::covariate1> > > acc3;
       auto sca = sourceCollectionForAnalysis.points();
       auto sc = secondaryCollection.points();
-      for (int i = 0; i < sourceCollectionForAnalysis.count(); i++) {
+      for (size_t i = 0; i < sourceCollectionForAnalysis.count(); i++) {
         Point p1 = sca.at(i);
         Point p2 = sc.at(i);
         acc1(p1.value);
